Split Android workspace settings package
