+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/CloudPostAuthRouteTest.kt
@@ -15,7 +15,7 @@ import com.flashcardsopensourceapp.data.local.model.CloudWorkspaceLinkSelection
 import com.flashcardsopensourceapp.feature.settings.cloud.CloudPostAuthMode
 import com.flashcardsopensourceapp.feature.settings.cloud.CloudPostAuthRoute
 import com.flashcardsopensourceapp.feature.settings.cloud.CloudPostAuthUiState
-import com.flashcardsopensourceapp.feature.settings.workspace.CurrentWorkspaceItemUiState
+import com.flashcardsopensourceapp.feature.settings.workspace.current.CurrentWorkspaceItemUiState
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Rule
