+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/LiveSmokeWorkspaceLifecycleFlows.kt
@@ -12,21 +12,21 @@ import androidx.compose.ui.test.performScrollToNode
 import androidx.compose.ui.test.performTextReplacement
 import com.flashcardsopensourceapp.app.FlashcardsApplication
 import com.flashcardsopensourceapp.data.local.model.CloudAccountState
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceCreateButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceExistingRowTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceListTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationDialogTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationErrorTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationFieldTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationLoadingTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteConfirmationPhraseTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeletePreviewBodyTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeletePreviewContinueButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeletePreviewDialogTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewDeleteWorkspaceButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewNameFieldTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewSaveNameButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceCreateButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceExistingRowTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceListTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteConfirmationButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteConfirmationDialogTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteConfirmationErrorTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteConfirmationFieldTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteConfirmationLoadingTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteConfirmationPhraseTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeletePreviewBodyTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeletePreviewContinueButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeletePreviewDialogTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewDeleteWorkspaceButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewNameFieldTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewSaveNameButtonTag
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
