+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/LiveSmokeWorkspaceProgressFlows.kt
@@ -7,18 +7,18 @@ import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performTextReplacement
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewTodayDueCountTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewTodayNewCountTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewTodayReviewedCountTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationDialogTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationFieldTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressConfirmationPhraseTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressErrorTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressPreviewBodyTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressPreviewButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceSettingsResetProgressPreviewDialogTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewTodayDueCountTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewTodayNewCountTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewTodayReviewedCountTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressConfirmationButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressConfirmationDialogTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressConfirmationFieldTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressConfirmationPhraseTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressErrorTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressPreviewBodyTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressPreviewButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.settings.workspaceSettingsResetProgressPreviewDialogTag
 
 internal fun LiveSmokeContext.resetWorkspaceProgressFromSettings(expectedCardsToResetCount: Int) {
     openSettingsSection(sectionTitle = "Workspace")
