+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/LiveSmokeWorkspaceStateSupport.kt
@@ -9,17 +9,17 @@ import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performScrollToNode
 import com.flashcardsopensourceapp.app.FlashcardsApplication
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceCreateButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceErrorMessageTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceExistingRowTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceListTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceLoadingStateTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceNameTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceOperationMessageTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceReloadButtonTag
-import com.flashcardsopensourceapp.feature.settings.workspace.currentWorkspaceSelectedSummaryTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewErrorMessageTag
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceOverviewNameFieldTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceCreateButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceErrorMessageTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceExistingRowTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceListTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceLoadingStateTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceNameTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceOperationMessageTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceReloadButtonTag
+import com.flashcardsopensourceapp.feature.settings.workspace.current.currentWorkspaceSelectedSummaryTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewErrorMessageTag
+import com.flashcardsopensourceapp.feature.settings.workspace.overview.workspaceOverviewNameFieldTag
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
