+++ apps/android/feature/settings/src/main/java/com/flashcardsopensourceapp/feature/settings/SettingsViewModel.kt
@@ -18,7 +18,7 @@ import com.flashcardsopensourceapp.data.local.repository.AutoSyncOutcome
 import com.flashcardsopensourceapp.data.local.repository.AutoSyncRequest
 import com.flashcardsopensourceapp.data.local.repository.CloudAccountRepository
 import com.flashcardsopensourceapp.data.local.repository.WorkspaceRepository
-import com.flashcardsopensourceapp.feature.settings.workspace.workspaceUpdatedOnAnotherDeviceMessage
+import com.flashcardsopensourceapp.feature.settings.workspace.shared.workspaceUpdatedOnAnotherDeviceMessage
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
