+++ apps/android/feature/settings/src/main/java/com/flashcardsopensourceapp/feature/settings/workspace/settings/WorkspaceScreenTags.kt
@@ -1,3 +1,3 @@
-package com.flashcardsopensourceapp.feature.settings.workspace
+package com.flashcardsopensourceapp.feature.settings.workspace.settings
 
 const val workspaceSchedulerSummaryTag: String = "workspace_scheduler_summary"
