+++ apps/android/feature/settings/src/main/java/com/flashcardsopensourceapp/feature/settings/workspace/shared/WorkspaceSyncSupport.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.feature.settings.workspace
+package com.flashcardsopensourceapp.feature.settings.workspace.shared
 
 import com.flashcardsopensourceapp.feature.settings.R
 import com.flashcardsopensourceapp.feature.settings.SettingsStringResolver
