+++ apps/android/feature/settings/src/main/java/com/flashcardsopensourceapp/feature/settings/workspace/tags/WorkspaceTagsRoute.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.feature.settings.workspace
+package com.flashcardsopensourceapp.feature.settings.workspace.tags
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
